Comic/ComicCrafter-AI/requirements.txt
streamlit  
pillow  
requests  
python-dotenv  
google-generativeai  
reportlab  
